refactor(video-preview): add a reusable useVideoPreview hook
Adds a new hook with reusable state and functions of video preview related functions so that they can be reused
in profile view and camera as content.
- Centralize core video preview code in one place
- Convert the code to typescript
- Remove duplicated code related to camera preview in the profile sidebar and share camera as content
